The BetaGameTheory Module

BetaGameTheory — Module

BetaGameTheory module

Provide utility functions to work with strategic games, including finding Hash equilibrium for n-player simultaneous games when mixed strategies are allowed.

Module Index

BetaGameTheory.plusTwo

Detailed API

BetaGameTheory.plusTwo — Method

plusTwo(x)

Sum the numeric "2" to whatever it receives as input

A more detailed explanation can go here, although I guess it is not needed in this case

Arguments

x: The amount to which we want to add 2

Notes

Notes can go here

Examples

julia> five = plusTwo(3)
5