The StrategicGames Module

StrategicGames — Module

StrategicGames module

Provide utility functions to work with strategic games, including finding Nash equilibrium simultaneous games when mixed strategies are allowed.

All functions work with generic n players (the examples generally show 2 players for simplicity)

Module Index

StrategicGames.best_response

StrategicGames.dominated_strategies

StrategicGames.expand_dimensions

StrategicGames.expected_value

StrategicGames.is_best_response

StrategicGames.is_nash

StrategicGames.nash_cp

Detailed API

StrategicGames.best_response — Method

best_response(payoff_array,strategy_profile,player)

Return (possibly one of many) best strategy and corrsponding expected payoff for a given player

Parameters:

payoff_array: the nplayers+1 array of payoffs

strategy_profile: the vector of vectors defining the strategies for the N players. The strategy for player n for which the best response is looked is used as init value in the optimisation

player: counter of the player for which we want to compute the best_response (e.g. 1 or 3)

Returns:

A named tuple with: expected_payoff, optimal_strategy, status (of the underlying optimisation)

Example:

`$julia julia> using StrategicGames julia> payoff_array = [(3,4) (1,5); (4,2) (2,3)] # prisoner's dilemma 2×2 Matrix{Tuple{Int64, Int64}}: (3, 4) (1, 5) (4, 2) (2, 3) julia> best_response(expand_dimensions(payoff_array),[[0.5,0.5],[0.5,0.5]],2) (expected_payoff = 4.0, optimal_strategy = [0.0, 1.0], status = MathOptInterface.OPTIMAL)$

StrategicGames.dominated_strategies — Method

dominated_strategies(payoff,player;strict=true)

Return a vector with the positions of the actions for player player that are dominates by at least one of his other actions.

Example

julia> using StrategicGames
julia> payoff_array = expand_dimensions([(3,4) (1,5) (6,2); (2,6) (3,7) (1,7)])
2×3×2 Array{Int64, 3}:
[:, :, 1] =
 3  1  6
 2  3  1
[:, :, 2] =
 4  5  2
 6  7  7
julia> dominated_strategies(payoff_array,2,strict=false) 
2-element Vector{Int64}:
 1
 3

StrategicGames.expand_dimensions — Method

expand_dimensions(x::AbstractArray{T}

Convenence function to transform a N dimensional array of tuples in a N+1 dimensional array of scalars (the format used for most functions in this package).

Example:

julia> using StrategicGames
julia> payoff_tuple = [(1,-1) (-1,1) (1,0); (-1,1) (1, -1) (0,1)] # 2 players, with 2 and 3 actions respectively
2×3 Matrix{Tuple{Int64, Int64}}:
 (1, -1)  (-1, 1)  (1, 0)
 (-1, 1)  (1, -1)  (0, 1)

julia> payoff_array = expand_dimensions(payoff_tuple)
2×3×2 Array{Int64, 3}:
[:, :, 1] =
  1  -1  1
 -1   1  0

[:, :, 2] =
 -1   1  0
  1  -1  1

StrategicGames.expected_value — Method

expected_value(v::Array{N,Real},p::Vector{Vector{Real}}) --> Real

Compute the expected value (scalar) of a N-dimensional value tensor given a vector of N probability vectors, one per each dimension of the value tensor.

StrategicGames.is_best_response — Method

is_best_response(payoff_array,strategy_profile,player;atol=1e-07,rtol=1e-07)

Determine if a given strategy for player player is the best response to a given strategy profile given a specific payoff matrix

Parameters:

payoff_array: the nplayers+1 array of payoffs

strategy_profile: the vector of vectors defining the strategies for the N players

player: counter of the player for which we want to verify if its strategy is a best_response (e.g. 1 or 3)

atol: absolute tollerance in comparing the expected utility from the given strategy and those from the optimal one [def: 1e-07]

rtol: relative tollerance in comparing the expected utility from the given strategy and those from the optimal one [def: 1e-07]

Example :

julia> using StrategicGames
julia> payoff_array = [(3,4) (1,5); (4,2) (2,3)] # prisoner's dilemma
2×2 Matrix{Tuple{Int64, Int64}}:
 (3, 4)  (1, 5)
 (4, 2)  (2, 3)
julia> is_best_response(expand_dimensions(payoff_array),[[0,1],[0.5,0.5]],1)
true
julia> is_best_response(expand_dimensions(payoff_array),[[0,1],[0.5,0.5]],2)
false

StrategicGames.is_nash — Method

is_nash(payoff_array,strategy_profile;atol=1e-07,rtol=1e-07)

Determine if a strategy profile is a Nash equilibrium for a given payoff matrix, i.e. all strategies are (weak) best responses

Parameters:

payoff_array: the nplayers+1 array of payoffs

strategy_profile: the vector of vectors defining the strategies for the N players

atol: absolute tollerance in comparing the expected utility from the given strategies and those from the optimal ones [def: 1e-07]

rtol: relative tollerance in comparing the expected utility from the given strategies and those from the optimal ones [def: 1e-07]

Example :

julia> using StrategicGames
julia> payoff_array  = [(3,4) (1,5); (4,2) (2,3)] # prisoner's dilemma
2×2 Matrix{Tuple{Int64, Int64}}:
 (3, 4)  (1, 5)
 (4, 2)  (2, 3)
julia> is_nash(expand_dimensions(payoff_array),[[0,1],[0,1]])
true

StrategicGames.nash_cp — Method

nash_cp(payoff_tensor;init,verbosity)

Find a Nash Equilibrium for n-players simultaneous games when mixed strategies are allowed using the Complementarity Problem algorithm.

Parameters

payoff_tensor: the nplayers+1 dimension payoff tensor of payoffs for the various players

init: a vector of vector of mixed strategies (i.e. PMFs) for each players to start the algorithm with. Different init points may reach different equilibrium points [def: equal probabilities for each available action by the players]

verbosity: an integer parameter to pass to the inner solver (currently Ipopt) [def: 0]

Notes

This function uses a LCP (Linear Complementarity) formulation (for nplayers >3 the algorithm actually is not linear)

The implementation uses the JuMP modelling language with the Ipopt solver engine (and hence it uses an interior point method instead of the pivotal approach used in the original Lemke-Howson [1964] algorithm)

There is no guarantee on timing and even that the algorithm converge to an equilibrium. Different equilibriums may be reached by setting different initial points

Returns

A named tuple with the following elements: status,equilibrium_strategies,expected_payoffs

Example

julia> using StrategicGames
julia> payoff = [(-1,-1) (-3,0); (0, -3) (-2, -2)] # prisoner's dilemma
2×2 Matrix{Tuple{Int64, Int64}}:
 (-1, -1)  (-3, 0)
 (0, -3)   (-2, -2)
julia> eq     = nash_cp(expand_dimensions(payoff));
julia> eq_strategies = eq.equilibrium_strategies
2-element Vector{Vector{Float64}}:
 [-4.049752569180346e-11, 1.0000000000404976]
 [-4.0497525691839856e-11, 1.0000000000404976]