The BetaGameTheory Module

BetaGameTheory — Module

BetaGameTheory module

Provide utility functions to work with strategic games, including finding Hash equilibrium for n-player simultaneous games when mixed strategies are allowed.

Module Index

BetaGameTheory.expand_dimensions

BetaGameTheory.expected_value

BetaGameTheory.nash_lcp

Detailed API

BetaGameTheory.expand_dimensions — Method

expand_dimensions(x::AbstractArray{T}

Convenence function to transform a N dimensional array of tuples in a N+1 dimensional array of scalars (the format used for most functions in this package).

Example:

julia> payoff_tuple = [(1,-1) (-1,1) (1,0); (-1,1) (1, -1) (0,1)] # 2 players, with 2 and 3 actions respectively
2×3 Matrix{Tuple{Int64, Int64}}:
 (1, -1)  (-1, 1)  (1, 0)
 (-1, 1)  (1, -1)  (0, 1)

julia> payoff_array = expand_dimensions(payoff_tuple)
2×3×2 Array{Int64, 3}:
[:, :, 1] =
  1  -1  1
 -1   1  0

[:, :, 2] =
 -1   1  0
  1  -1  1

BetaGameTheory.expected_value — Method

expected_value(v::Array{N,Real},p::Vector{Vector{Real}}) --> Real

Compute the expected value (scalar) of a N-dimensional value tensor given a vector of N probability vectors, one per each dimension of the value tensor.

BetaGameTheory.nash_lcp — Method

nash_lcp(payoff_tensor;init,verbosity)

Find a Nash Equilibrium for n-players simultaneous games when mixed strategies are allowed using the LCP algorithm.

Parameters

payoff_tensor: the nplayers+1 dimension payoff tensor of payoffs for the various players

init: a vector of vector of mixed strategies (i.e. PMFs) for each players to start the algorithm with. Different init points may reach different equilibrium points [def: equal probabilities for each available action by the players]

verbosity: an integer parameter to pass to the inner solver (currently Ipopt) [def: 0]

Notes

This funcion uses the LCP (Linear Complementarity) formulation from Lemke-Howson 1964

The implementation uses the JuMP modelling language with the Ipopt solver engine

There is no guarantee on timing and even that the algorithm converge to an equilibrium. Different equilibriums may be reached by setting different initial points

Returns

A named tuple with the following elements: status,equilibrium_strategies,expected_payoffs

Example